requirements.txt
crewai
python-dotenv
openai
streamlit
python-docx 
